Implements alternative needle calibration
Removes need for camera calibration step. New procedure makes use of an
inverted needle fixed in the field-of-view of the web cams. The gantry
head camera is brought to the inverted needle and the position is
recorded. Then, the loaded needle tip is brought into contact with the
inverted needle with the aide of the web-cams. The difference between this
position and the previous position is the needle-tip offset.